refactor: colocate completion tools with tracker effects
