+++ tests/infra/test_schema_version_convention.py
@@ -130,8 +130,8 @@ _LEGACY_JSON_WRITES: set[tuple[str, int]] = {
     ("src/autoskillit/server/tools/_overlay_state.py", 114),
     # tools_pipeline_tracker.py — tracker_data dict (init) and mark_step_complete write
     # (same tracker file schema as init — not a new format, grandfathered alongside it)
-    ("src/autoskillit/server/tools/tools_pipeline_tracker.py", 268),
-    ("src/autoskillit/server/tools/tools_pipeline_tracker.py", 478),
+    ("src/autoskillit/server/tools/tools_pipeline_tracker.py", 273),
+    ("src/autoskillit/server/tools/tools_pipeline_tracker.py", 483),
     # tools_status.py — mcp_data dict
     ("src/autoskillit/server/tools/tools_status.py", 536),
     # tools_github.py — bug report dict (non-blocking report-bug status file)
