+++ tests/integration/test_pipeline_step_completion_flow.py
@@ -17,7 +17,7 @@ from autoskillit.core.types import RetryReason
 from autoskillit.core.types._type_results import SkillResult
 from autoskillit.server._run_skill_completion import RunSkillCompletionMiddleware
 from autoskillit.server.tools.tools_execution import run_skill
-from autoskillit.server.tools.tools_run_skill_completion import (
+from autoskillit.server.tools.tools_pipeline_tracker import (
     complete_run_skill_result,
     recover_run_skill_result,
 )
